move csv parseing functions to /util
