+++ src/util/csv.ts
@@ -1,0 +1,29 @@
+const splitWhenNotInQuotes = (input: string, delimiter: string): string[] => {
+	let pos = 0;
+	let inQuotes = false;
+	const output = [];
+	let part = '';
+
+	for (pos = 0; pos < input.length; pos++) {
+		const c = input[pos];
+		if (c === '"') {
+			inQuotes = !inQuotes;
+		}
+		if (c === delimiter && !inQuotes) {
+			output.push(part);
+			part = '';
+		} else {
+			part += c;
+		}
+	}
+
+	return output;
+};
+
+export const parseStringifiedCSV = (csvData: string): string[][] => {
+	const rows = splitWhenNotInQuotes(csvData, '\n').filter(
+		(_) => _.length > 0,
+	);
+	const cellsInRows = rows.map((_) => splitWhenNotInQuotes(_, ','));
+	return cellsInRows;
+};
